update geoda wasm
