+++ .github/workflows/nextjs.yml
@@ -29,6 +29,11 @@ jobs:
         uses: actions/setup-node@v3
         with:
           node-version: '18'
+      - name: Setup Emscripten
+        uses: mymindstorm/setup-emsdk@v11
+        with:
+          version: 3.1.34
+          actions-cache-folder: 'emsdk-cache'
       - name: Setup Pages
         uses: actions/configure-pages@v3
         with:
@@ -39,6 +44,7 @@ jobs:
         with:
           path: |
             /home/runner/work/reactgeoda/csds_kepler
+            /home/runner/work/reactgeoda/geoda-lib
           key: ${{ runner.os }}-csds_kepler-${{ hashFiles('deps.txt') }}
       - name: Install dependencies
         if: steps.cache-dependencies.outputs.cache-hit != 'true'
@@ -51,6 +57,12 @@ jobs:
           git clone https://github.com/GeoDaCenter/kepler.gl.git --branch=xli/reactgeoda csds_kepler
           cd csds_kepler
           yarn
+          cd ..
+          git clone https://github.com/GeoDaCenter/geoda-lib.git geoda-lib
+          cd geoda-lib
+          cd js
+          yarn
+          yarn build 
       - name: Run Lint
         run: |
           cd geoda-ai
