+++ geoda-ai/next.config.js
@@ -28,7 +28,7 @@ const nextConfig = {
     config.resolve.alias = {
       ...config.resolve.alias,
       'apache-arrow': resolve(__dirname, './node_modules/apache-arrow'),
-      'geoda-wasm': resolve(__dirname, '../../geoda-wasm/js'),
+      'geoda-wasm': resolve(__dirname, '../../geoda-lib/js'),
       // '@dnd-kit/core': resolve(__dirname, '../node_modules/@dnd-kit/core'),
       '@mapbox/tiny-sdf': resolve(
         __dirname,
@@ -91,8 +91,8 @@ const nextConfig = {
       new CopyPlugin({
         patterns: [
           {
-            from: resolve(__dirname, '../../geoda-wasm/js/dist/geoda.wasm'),
-            to: 'public/_next/static/chunks'
+            from: resolve(__dirname, '../../geoda-lib/js/dist/geoda.wasm'),
+            to: 'public/geoda.wasm'
           }
         ]
       })
