+++ geoda-ai/src/styles/settings-panel.css
@@ -75,14 +75,12 @@
 }
 
 .settings-panel .form-wrapper {
-  align-items: center;
   align-self: stretch;
   display: flex;
   flex: 1;
   flex-direction: column;
   flex-grow: 1;
-  gap: 20px;
-  padding: 20px 16px 0px;
+  padding: 16px 16px 0px;
   position: relative;
   width: 100%;
 }
