+++ geoda-ai/src/styles/style.css
@@ -232,14 +232,18 @@ div[mapboxgl-children] {
 }
 .cs-message-list {
   background-color: transparent!important;
+  margin-bottom: 20px!important;
 }
 .scrollbar-container {
-  padding: 0;
+  padding: 0!important;
 }
 .cs-message-input {
   border: none!important;
   background-color: transparent!important;
-  padding-bottom: 2px;
+  padding-bottom: 12px!important;
+}
+.cs-message--outgoing {
+  padding-right: 12px!important;
 }
 .cs-message--outgoing .cs-message__content {
   color: #ffffff !important;
@@ -247,4 +251,7 @@ div[mapboxgl-children] {
 }
 .cs-message__html-content {
   font-size: var(--text-sm-regular-font-size);
+}
+.cs-message-list__scroll-wrapper {
+  padding: 0!important;
 }
